requirements.txt
fastai
flask
gunicorn
google-cloud-storage

